Move new ITs into separate module and record coverage
