+++ integration-tests/oauth2/src/test/groovy/com/okta/spring/tests/oauth2/code/BasicRedirectCodeFlowApplication.groovy
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2017 Okta, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License")
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.okta.spring.tests.oauth2.code
+
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.autoconfigure.security.oauth2.client.EnableOAuth2Sso
+import org.springframework.security.access.expression.method.MethodSecurityExpressionHandler
+import org.springframework.security.access.prepost.PreAuthorize
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity
+import org.springframework.security.config.annotation.method.configuration.GlobalMethodSecurityConfiguration
+import org.springframework.security.oauth2.provider.expression.OAuth2MethodSecurityExpressionHandler
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.RestController
+
+import java.security.Principal
+
+@SpringBootApplication
+@RestController
+@EnableOAuth2Sso
+class BasicRedirectCodeFlowApplication {
+
+    @EnableGlobalMethodSecurity(prePostEnabled = true)
+    protected static class GlobalSecurityConfiguration extends GlobalMethodSecurityConfiguration {
+        @Override
+        protected MethodSecurityExpressionHandler createExpressionHandler() {
+            return new OAuth2MethodSecurityExpressionHandler()
+        }
+    }
+
+    @GetMapping("/")
+    @PreAuthorize("#oauth2.hasScope('email')")
+    String getMessageOfTheDay(Principal principal) {
+        return "The message of the day is boring: " + principal.getName()
+    }
+
+    static void main(String[] args) {
+        SpringApplication.run(BasicRedirectCodeFlowApplication.class, args)
+    }
+}
