+++ integration-tests/oauth2/src/test/groovy/com/okta/spring/tests/wiremock/QuickHttpMockTest.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.okta.spring.example.wiremock
+package com.okta.spring.tests.wiremock
 
 import com.github.tomakehurst.wiremock.WireMockServer
 import com.okta.spring.oauth.discovery.OidcDiscoveryClient
