+++ integration-tests/oauth2/src/test/resources/stubs/__files/discovery.json
