+++ integration-tests/oauth2/src/test/resources/stubs/__files/remote-validation-token.json
