+++ pom.xml
@@ -39,6 +39,7 @@
         <module>config</module>
         <module>oauth2</module>
         <module>okta-spring-boot-starter</module>
+        <module>integration-tests/oauth2</module>
         <module>examples</module>
         <module>coverage</module>
     </modules>
@@ -94,6 +95,11 @@
             <dependency>
                 <groupId>com.okta.spring</groupId>
                 <artifactId>okta-spring-config</artifactId>
+                <version>0.2.0-SNAPSHOT</version>
+            </dependency>
+            <dependency>
+                <groupId>com.okta.spring</groupId>
+                <artifactId>okta-spring-boot-integration-tests-oauth2</artifactId>
                 <version>0.2.0-SNAPSHOT</version>
             </dependency>
